v0.6
Started working on actual Village / Hub Scene - 30 % Done

Tried to play with Ground Shader Graph and Lighting Data

Tried to add Hieght map and Normal map to ground to give it depth only to know that, what I want is displacement and tesselation, only to know that it is only in HDRP and mine is URP, tried to convert it to HDRP, which is not possible for WebGL's so I reverted it, and only added Normal Maps

Added, Water, Foilage, and Rocks to Village / Hub Scene

Village / Hub - 90% Done

Added Skybox

When clicking the players house, the camera will transition inside the house with other options such as change inventory, change equipment and saving the game,

When clicking the players house, the camera will transition inside the house with other options such as change inventory, change equipment and saving the game,

Continued scripting and designing the change equipment function, it now outputs Currently Equipped Weapon and shows Data Values of:: Name, Type, Damage, Icons, (Elements) and a preview to compare hovered weapons with, showing the same data values and comparing them, Damage Text changes colors depending on when it is lower higher or neutral

Created appropriate graphics for weapon icons (Light club and Heavy Club)

Continued working on the Equipment function, added and Text E to show which weapon is currently equipped on the player,

also started studying on how to save and store data a list of weapons to my computer, i wanted to start working on the save function now while I only have one variable at the time, so when the time comes that I need to save multiple variables I won;t be overwhelmed. spent the whole day trying to understand how it works, and prepared the input and output values to store and load for my equipment manager for next week.

Couple of Scripts for my save function

Made a lot of script and deleting old ones relating to saving

FINALLY was able to make a system that works using a local database (Scitpable Object) that assigns IDs to a list of my weapons and a class that reads and determines what weapon it is pertaining to that ID, then I save the list integers of IDs from my weapons and covert it into binary, my game can now store and load values to the local machine successfully.

I was also able to continue with the rest of the project by making an list of armor that uses the same system as my weapons, all the prototype UI and scripts also works with it as well